punto 1 funcionando, api rest

+++ src/main/java/com/maq/challenge/service/DinosaurService.java
@@ -2,16 +2,21 @@ package com.maq.challenge.service;
 
 import com.maq.challenge.domain.Dinosaur;
 import com.maq.challenge.domain.DinoStatus;
+import com.maq.challenge.exception.DinosaurNotFoundException;
 import com.maq.challenge.persistence.DinosaurDE;
 import com.maq.challenge.persistence.mappers.DinosaurMapper;
 import com.maq.challenge.persistence.repository.DinosaurRepository;
 import com.maq.challenge.exception.DinosaurException;
+import lombok.extern.log4j.Log4j2;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.sql.SQLIntegrityConstraintViolationException;
 import java.util.List;
 
+@Log4j2
 @Service
 @Transactional
 public class DinosaurService {
@@ -35,20 +40,26 @@ public class DinosaurService {
 
     public Dinosaur update(Long id, Dinosaur dinosaur) {
         DinosaurDE existing = repository.findById(id)
-                .orElseThrow(() -> new DinosaurException("El Dinosaurio no existe"));
+                .orElseThrow(() -> new DinosaurNotFoundException("El Dinosaurio no existe"));
 
         if (existing.getStatus() == DinoStatus.EXTINCT) {
             throw new DinosaurException("No se puede actualizar un dinosaurio extinto");
         }
 
         validateDinosaur(dinosaur);
-        existing.setName(dinosaur.getName());
-        existing.setSpecies(dinosaur.getSpecies());
-        existing.setDiscoveryDate(dinosaur.getDiscoveryDate());
-        existing.setExtinctionDate(dinosaur.getExtinctionDate());
-        existing.setStatus(dinosaur.getStatus());
 
-        return mapper.toDomain(repository.save(existing));
+        if (dinosaur.getName() != null) existing.setName(dinosaur.getName());
+        if (dinosaur.getSpecies() != null) existing.setSpecies(dinosaur.getSpecies());
+        if (dinosaur.getDiscoveryDate() != null) existing.setDiscoveryDate(dinosaur.getDiscoveryDate());
+        if (dinosaur.getExtinctionDate() != null) existing.setExtinctionDate(dinosaur.getExtinctionDate());
+        if (dinosaur.getStatus() != null) existing.setStatus(dinosaur.getStatus());
+
+        try {
+            return mapper.toDomain(repository.save(existing));
+        } catch (DataIntegrityViolationException e) {
+            //tambien puedo usar el repository.existsByName(dinosaur.getName()) pero implica una query extra.
+            throw new DinosaurException("Ya existe un dinosaurio con ese nombre");
+        }
     }
 
     private void validateDinosaur(Dinosaur dinosaur) {
@@ -59,6 +70,9 @@ public class DinosaurService {
     }
 
     public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new DinosaurNotFoundException("Dinosaur does not exist");
+        }
         repository.deleteById(id);
     }
 
@@ -71,6 +85,6 @@ public class DinosaurService {
     }
 
     public Dinosaur findById(Long id) {
-        return mapper.toDomain(repository.findById(id).orElseThrow(() -> new DinosaurException("El Dinosaurio no existe")));
+        return mapper.toDomain(repository.findById(id).orElseThrow(() -> new DinosaurNotFoundException("El Dinosaurio no existe")));
     }
 }
